+++ lading/src/generator/tcp_rr.rs
@@ -5,6 +5,9 @@
 //! waits for a fixed-size response, and repeats. Flows are distributed across
 //! OS threads and multiplexed via mio.
 //!
+//! The event-loop machinery lives in [`crate::neper::rr`]; this module is a
+//! thin wrapper that supplies configuration.
+//!
 //! ## Metrics
 //!
 //! `requests_sent`: Completed request writes
@@ -13,23 +16,14 @@
 //! `bytes_read`: Response bytes received
 //! `connections_failed`: Failed connection attempts
 
-use std::io::{self, ErrorKind, Read, Write};
-use std::net::{self, IpAddr, SocketAddr};
+use std::net::{IpAddr, SocketAddr};
 use std::num::{NonZeroU16, NonZeroUsize};
-use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, Ordering::Relaxed};
-use std::time::{Duration, Instant};
 
-use mio::net::TcpStream;
-use mio::{Events, Interest, Poll, Token};
 use serde::{Deserialize, Serialize};
-use tracing::{info, trace};
 
 use super::General;
 use crate::generator::common::MetricsBuilder;
-use crate::neper::flow::{self, Action, Flow, FlowMap};
-use crate::neper::metrics::{self, ThreadMetrics};
-use crate::neper::thread;
+use crate::neper::rr::{self, ClientParams};
 
 fn default_nonzero_u16() -> NonZeroU16 {
     NonZeroU16::new(1).expect("1 is nonzero")
@@ -83,15 +77,9 @@ pub struct Config {
 #[derive(thiserror::Error, Debug)]
 /// Errors produced by [`TcpRr`].
 pub enum Error {
-    /// IO error
+    /// Shared neper-style request/response error.
     #[error(transparent)]
-    Io(#[from] std::io::Error),
-    /// Worker thread panicked
-    #[error("Worker thread panicked")]
-    ThreadPanicked,
-    /// Invalid configuration.
-    #[error("invalid config: {0}")]
-    Config(String),
+    Rr(#[from] rr::Error),
 }
 
 #[derive(Debug)]
@@ -100,11 +88,6 @@ pub struct TcpRr {
     config: Config,
     metric_labels: Vec<(String, String)>,
     shutdown: lading_signal::Watcher,
-}
-
-enum ClientState {
-    SendRequest,
-    RecvResponse,
 }
 
 impl TcpRr {
@@ -123,226 +106,23 @@ impl TcpRr {
     ///
     /// # Errors
     ///
-    /// Returns an error if a worker thread panics.
+    /// Returns an error if a worker thread panics or configuration is invalid.
     ///
     /// # Panics
     ///
-    /// Panics if `addr` cannot be resolved to a socket address.
+    /// Panics if `addr` cannot be parsed as an IP address.
     pub async fn spin(self) -> Result<(), Error> {
-        if self.config.threads > self.config.flows {
-            return Err(Error::Config(format!(
-                "threads ({}) must be <= flows ({})",
-                self.config.threads, self.config.flows
-            )));
-        }
-
         let ip: IpAddr = self.config.addr.parse().expect("invalid addr");
-        let data_addr = SocketAddr::new(ip, self.config.data_port);
-        let control_addr = SocketAddr::new(ip, self.config.control_port);
-
-        let shutdown_flag = thread::new_shutdown_flag();
-
-        // Wait for the blackhole to be ready by connecting to its control port.
-        info!("waiting for blackhole control port at {control_addr}");
-        let deadline = Instant::now() + Duration::from_secs(300);
-        {
-            let flag = Arc::clone(&shutdown_flag);
-            let shutdown = self.shutdown.clone();
-            tokio::spawn(async move {
-                shutdown.recv().await;
-                flag.store(true, Relaxed);
-            });
-        }
-        loop {
-            if shutdown_flag.load(Relaxed) {
-                return Err(Error::Io(io::Error::new(
-                    ErrorKind::ConnectionRefused,
-                    format!(
-                        "shutdown before blackhole control port {control_addr} became reachable"
-                    ),
-                )));
-            }
-            match net::TcpStream::connect(control_addr) {
-                Ok(_conn) => {
-                    info!("blackhole ready, starting flows");
-                    break;
-                }
-                Err(e) => {
-                    if Instant::now() >= deadline {
-                        return Err(Error::Io(io::Error::new(
-                            ErrorKind::TimedOut,
-                            format!(
-                                "blackhole control port {control_addr} not reachable after 5 minutes: {e}"
-                            ),
-                        )));
-                    }
-                    std::thread::sleep(Duration::from_millis(100));
-                }
-            }
-        }
-        let num_threads = self.config.threads.get();
-        let num_flows = self.config.flows.get();
-        let request_size = self.config.request_size.get();
-        let response_size = self.config.response_size.get();
-
-        let flow_dist = thread::distribute_flows(num_flows, num_threads);
-
-        let thread_metrics = Arc::new(
-            (0..num_threads)
-                .map(|_| ThreadMetrics::new())
-                .collect::<Vec<_>>(),
-        );
-
-        let metrics_handle = {
-            let tm = Arc::clone(&thread_metrics);
-            let labels = self.metric_labels.clone();
-            let flag = Arc::clone(&shutdown_flag);
-            thread::spawn_named("tcp_rr-metrics", move || {
-                metrics::run_metrics_thread(&tm, &labels, &flag);
-            })
+        let params = ClientParams {
+            data_addr: SocketAddr::new(ip, self.config.data_port),
+            control_addr: SocketAddr::new(ip, self.config.control_port),
+            threads: self.config.threads.get(),
+            flows: self.config.flows.get(),
+            request_size: self.config.request_size.get(),
+            response_size: self.config.response_size.get(),
+            no_delay: self.config.no_delay,
         };
-
-        let mut worker_handles = Vec::with_capacity(num_threads as usize);
-        for i in 0..num_threads {
-            let thread_flows = flow_dist[i as usize];
-            let flag = Arc::clone(&shutdown_flag);
-            let tm = Arc::clone(&thread_metrics);
-            let no_delay = self.config.no_delay;
-            let handle = thread::spawn_named(&format!("tcp_rr-client-{i}"), move || {
-                client_thread_main(
-                    data_addr,
-                    thread_flows,
-                    request_size,
-                    response_size,
-                    no_delay,
-                    &flag,
-                    &tm[i as usize],
-                );
-            });
-            worker_handles.push(handle);
-        }
-
-        self.shutdown.recv().await;
-        info!("shutdown signal received");
-        shutdown_flag.store(true, Relaxed);
-
-        worker_handles.push(metrics_handle);
-        thread::join_all(worker_handles).map_err(|()| Error::ThreadPanicked)?;
-
+        rr::run_client(params, self.metric_labels, self.shutdown, "tcp_rr").await?;
         Ok(())
-    }
-}
-
-fn client_thread_main(
-    addr: SocketAddr,
-    num_flows: u16,
-    request_size: usize,
-    response_size: usize,
-    no_delay: bool,
-    shutdown_flag: &AtomicBool,
-    metrics: &ThreadMetrics,
-) {
-    let mut poll = Poll::new().expect("failed to create mio::Poll");
-    let mut events = Events::with_capacity(num_flows as usize);
-    let request_buf = vec![0u8; request_size];
-    let mut response_buf = vec![0u8; response_size];
-    let mut flows: FlowMap<ClientState> = FlowMap::new();
-    let mut next_token: usize = 0;
-
-    for _ in 0..num_flows {
-        match net::TcpStream::connect(addr) {
-            Ok(std_stream) => {
-                let _ = std_stream.set_nodelay(no_delay);
-                std_stream
-                    .set_nonblocking(true)
-                    .expect("failed to set nonblocking");
-                let mut stream = TcpStream::from_std(std_stream);
-                let token = Token(next_token);
-                next_token += 1;
-                poll.registry()
-                    .register(&mut stream, token, Interest::WRITABLE)
-                    .expect("failed to register flow");
-                flows.insert(Flow {
-                    stream,
-                    token,
-                    state: ClientState::SendRequest,
-                    xfer: request_size,
-                });
-            }
-            Err(e) => {
-                trace!("connection to {addr} failed: {e}");
-                metrics.connections_failed.add(1);
-            }
-        }
-    }
-
-    loop {
-        let _ = poll.poll(&mut events, Some(Duration::from_millis(100)));
-        if shutdown_flag.load(Relaxed) {
-            break;
-        }
-        for event in &events {
-            let token = event.token();
-            let Some(fl) = flows.get_mut(token) else {
-                continue;
-            };
-            let action = handle_client_event(fl, &request_buf, &mut response_buf, metrics);
-            flow::apply_action(action, token, &mut flows, poll.registry());
-        }
-    }
-}
-
-fn handle_client_event(
-    flow: &mut Flow<ClientState>,
-    request_buf: &[u8],
-    response_buf: &mut [u8],
-    metrics: &ThreadMetrics,
-) -> Action {
-    match flow.state {
-        ClientState::SendRequest => {
-            let offset = request_buf.len() - flow.xfer;
-            match flow.stream.write(&request_buf[offset..]) {
-                Ok(n) => {
-                    flow.xfer -= n;
-                    if flow.xfer == 0 {
-                        flow.xfer = response_buf.len();
-                        flow.state = ClientState::RecvResponse;
-                        metrics.requests_sent.add(1);
-                        metrics.bytes_written.add(request_buf.len() as u64);
-                        Action::Reregister(Interest::READABLE)
-                    } else {
-                        Action::Continue
-                    }
-                }
-                Err(e) if e.kind() == ErrorKind::WouldBlock => Action::Continue,
-                Err(e) => {
-                    trace!("write error: {e}");
-                    Action::Remove
-                }
-            }
-        }
-        ClientState::RecvResponse => {
-            let offset = response_buf.len() - flow.xfer;
-            match flow.stream.read(&mut response_buf[offset..]) {
-                Ok(0) => Action::Remove,
-                Ok(n) => {
-                    flow.xfer -= n;
-                    if flow.xfer == 0 {
-                        flow.xfer = request_buf.len();
-                        flow.state = ClientState::SendRequest;
-                        metrics.responses_received.add(1);
-                        metrics.bytes_read.add(response_buf.len() as u64);
-                        Action::Reregister(Interest::WRITABLE)
-                    } else {
-                        Action::Continue
-                    }
-                }
-                Err(e) if e.kind() == ErrorKind::WouldBlock => Action::Continue,
-                Err(e) => {
-                    trace!("read error: {e}");
-                    Action::Remove
-                }
-            }
-        }
     }
 }
