+++ lading/src/neper/metrics.rs
@@ -69,8 +69,10 @@ define_thread_metrics! {
     responses_received,
     bytes_written,
     bytes_read,
+    connections_initiated,
     connections_failed,
     connections_accepted,
+    connections_closed,
     requests_received,
     responses_sent,
     bytes_received,
